Allow llm_job_func to report to mission-specific report delegate
