+++ pipelex/cogt/openai/openai_llm_worker.py
@@ -14,9 +14,10 @@ from pipelex import log
 from pipelex.cogt.exceptions import SdkTypeError
 from pipelex.cogt.inference.inference_report_delegate import InferenceReportDelegate
 from pipelex.cogt.llm.llm_job import LLMJob
+from pipelex.cogt.llm.llm_job_func import llm_job_func
 from pipelex.cogt.llm.llm_models.llm_engine import LLMEngine
 from pipelex.cogt.llm.llm_models.llm_family import LLMFamily
-from pipelex.cogt.llm.llm_worker_abstract import LLMWorkerAbstract, llm_job_func
+from pipelex.cogt.llm.llm_worker_abstract import LLMWorkerAbstract
 from pipelex.cogt.llm.structured_output import StructureMethod
 from pipelex.cogt.openai.openai_errors import OpenAIWorkerError
 from pipelex.cogt.openai.openai_factory import OpenAIFactory
